Add CentOS 7 to 8 Script

+++ voxscript.sh
@@ -12,6 +12,7 @@ if "$lang" == "en":
 	echo "Nb Script: 2"
 	echo "1- Update Debian 8 to 10 (deb8to10.sh)"
 	echo "2- Install flood seedbox in Debian 10 (seedbox_deb10_flood.sh)"
+	echo "3- Update CentOS 7 to 8 (centos7to8.sh)"
 	read -p "What's your choices ?" chx
 	if "$chx" == "1":
 		mkdir /tmp/vscript
@@ -27,6 +28,13 @@ if "$lang" == "en":
 		bash /tmp/vscript/seedbox_deb10_flood.sh
 		rm -r /tmp/vscript
 		exit
+	if "$chx" == "3":
+		mkdir /tmp/vscript
+		curl https://voxan24.github.io/Linux_Script/centos7to8.sh -o /tmp/vscript/centos7to8.sh
+		chmod a+x /tmp/vscript/centos7to8.sh
+		bash /tmp/vscript/centos7to8.sh
+		rm -r /tmp/vscript
+		exit
 	else:
 		echo "Invalid, goodbye"
 		exit
@@ -36,6 +44,7 @@ if "lang" == "fr":
 	echo "Nombre de script disponible: 2"
 	echo "1- Script de mise a niveaux de Debian 8 vers Debian 10 (deb8to10.sh)"
 	echo "2- Script d'installation d'une seedbox avec flood sous Debian 10 (seedbox_deb10_flood.sh)"
+	echo "3- Script de mise a niveaux de CentOS 7 vers CentOS 8 (centos7to8.sh)"
 	read -p "Quel est votre choix (entrez le chiffre correspondant a votre choix)" chx
 	if "$chx" == "1":
 		mkdir /tmp/vscript
@@ -49,6 +58,13 @@ if "lang" == "fr":
 		curl https://voxan24.github.io/Linux_Script/seedbox_deb10_flood.sh -o /tmp/vscript/seedbox_deb10_flood.sh
 		chmod a+x /tmp/vscript/seedbox_deb10_flood.sh
 		bash /tmp/vscript/seedbox_deb10_flood.sh
+		rm -r /tmp/vscript
+		exit
+	if "$chx" == "3":
+		mkdir /tmp/vscript
+		curl https://voxan24.github.io/Linux_Script/centos7to8.sh -o /tmp/vscript/centos7to8.sh
+		chmod a+x /tmp/vscript/centos7to8.sh
+		bash /tmp/vscript/centos7to8.sh
 		rm -r /tmp/vscript
 		exit
 	else:
